fixed UserController Error

+++ Backend/springboot_manytoone/src/main/java/manytoone/Groups/UserGroupRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserGroupRepository extends JpaRepository<UserGroup, Integer> {
+public interface UserGroupRepository extends JpaRepository<UserGroup, Long> {
     List<UserGroup> findByUserId(int userId);
     List<UserGroup> findByGroupId(int groupId);
     Optional<UserGroup> findByUserIdAndGroupId(int userId, int groupId);
